Define mindeps API - allow calling `entrypoint` and `get_min_deps`
